WP-PF.2 settle: honor the flicker Double-arithmetic model on window controls
- flicker.rs + post.rs: compute `trunc(600/ts)` and `(t-tPst)>=600` in f64
  over f32 operands (Delphi Win64 Extended≡Double), matching the module's
  stated model; makes `hst` sizing equal Delphi's `SetLength` exactly (probe:
  ts=0.1 -> 6000 not 6001). Golden stays bit-exact; subtraction is Sterbenz-exact.
- monitor accessors: correct the do_action doc — Process/Save/Take need circuit/
  solution/file context the parse hook lacks (post-process runs via Export/Show).
- add a fractional-step (ts=0.1) flicker unit test for the previously untested regime.

+++ crates/dss-core/src/elements/meter/monitor/post.rs
@@ -56,7 +56,10 @@ impl Monitor {
         }
 
         // Npst = 1 + Trunc(t_N / 600); Vbase = 1000·kVBase (narrowed to Single).
-        let npst = 1 + (times[n - 1] / 600.0).trunc() as usize;
+        // Pascal `Trunc(data[0][N] / 600.0)`: `data[0][N]` is `Single`, `600.0` a
+        // `Double` literal, so the division is `Double` (Delphi Win64) — matches
+        // `flicker::flicker_meter`'s Single-storage/Double-arithmetic model.
+        let npst = 1 + (times[n - 1] as f64 / 600.0).trunc() as usize;
         let vbase = ((1000.0 * kv_base) as f32) as f64;
         let f_base = self.med.cd.base_frequency;
 
@@ -73,7 +76,10 @@ impl Monitor {
         let mut t_pst = 0.0_f32;
         let mut ipst = 0usize;
         for i in 0..n {
-            if (times[i] - t_pst) >= 600.0 {
+            // Pascal `(data[0][i] - tpst) >= 600.0`: `Single` operands promoted to
+            // `Double` for the subtraction/compare (Delphi Win64); `t_pst` stays
+            // f32-stored, as Pascal `tpst: Single`.
+            if (times[i] as f64 - t_pst as f64) >= 600.0 {
                 ipst += 1;
                 t_pst = times[i];
             }
